pyproject.toml
[tool.poetry]
name = "superforge_poetry"
version = "0.1.0"
description = "Automate the scraping, editing, and formatting of the web novel, Super Gene, into ePub ebooks."
authors = ["Max Ludden <dev@maxludden.com>"]

[tool.poetry.dependencies]
python = "^3.10"
mongoengine = "^0.24.2"
loguru = "^0.6.0"
tqdm = "^4.64.0"
alive-progress = "^2.4.1"
markdown2 = "^2.4.3"
sh = "^1.14.3"
PyYAML = "^6.0"
requests = "^2.28.1"
bs4 = "^0.0.1"
selenium = "^4.4.0"
python-dotenv = "^0.20.0"
num2words = "^0.5.11"
rich = {extras = ["all"], version = "^12.5.1"}
pathlib-tree = "^2.2.1"
beanie = "^1.11.9"
ujson = "^5.4.0"
ipython = "^8.4.0"
ipywidgets = "^8.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
get_logger = "max_logger.main:get_logger"
